Feature: Organizacion de carpetas y primer acercamiento a la conexion de la base de datos
