--- app.vue
@@ -1,9 +1,0 @@
-<template>
-<div>
-  <NuxtRouteAnnouncer /> 
-  
-  <NuxtLayout>
-    <NuxtPage />
-  </NuxtLayout>
-</div>
-</template>
