--- layouts/default.vue
@@ -1,9 +1,0 @@
-<template>
-  <div class="min-h-screen bg-gray-100 font-sans">
-    <Navbar />
-    <main>
-      <slot />
-    </main>
-    <Footer />
-  </div>
-</template>
