+++ nuxt.config.ts
@@ -1,5 +1,6 @@
 // https://nuxt.com/docs/api/configuration/nuxt-config
 export default defineNuxtConfig({
+  srcDir: "app/",
   compatibilityDate: '2024-04-03',
   modules: ['@nuxtjs/tailwindcss'],
   devtools: { enabled: false },
